+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 #### New Features
 
-* Added support for circular dependencies by adding `ObjectGraph` scope to reuse resolved instances.  
+* Added support for circular dependencies:
+    * Added `ObjectGraph` scope to reuse resolved instances 
+    * Added `resolveDependencies` method on `DefinitionOf` class to resolve dependencies of resolved instance.  
   [#11](https://github.com/AliSoftware/Dip/pull/11), [@ilyapuchka](https://github.com/ilyapuchka)
 * Added methods to register/remove individual definitions.  
   [#11](https://github.com/AliSoftware/Dip/pull/11), [@ilyapuchka](https://github.com/ilyapuchka)