+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Develop
+## 4.0.0
 
 #### New Features